test(server): colocate stage router stats coverage
